+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - use commit and push logic for changelog update(pr [#301])
+- add Rebase command to CLI(pr [#304])
 
 ### Changed
 
@@ -579,6 +580,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#301]: https://github.com/jerus-org/pcu/pull/301
 [#302]: https://github.com/jerus-org/pcu/pull/302
 [#303]: https://github.com/jerus-org/pcu/pull/303
+[#304]: https://github.com/jerus-org/pcu/pull/304
 [Unreleased]: https://github.com/jerus-org/pcu/compare/v0.4.1...HEAD
 [0.4.1]: https://github.com/jerus-org/pcu/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/jerus-org/pcu/compare/v0.3.0...v0.4.0